+++ app/web/requests/app_request_test.rb
@@ -1,0 +1,7 @@
+class AppRequestTest < Liza::RequestTest
+
+  test :subject_class do
+    assert subject_class == AppRequest
+  end
+
+end