+++ app/web_box.rb
@@ -1,0 +1,8 @@
+class WebBox < Liza::WebBox
+
+  requests do
+    # Set up your request panel per the DSL in http://guides.lizarb.org/panels/request.html
+
+  end
+
+end