+++ lib/web_system.rb
@@ -1,0 +1,5 @@
+require "uri"
+
+class WebSystem < Liza::System
+  set :log_color, :blue
+end