+++ lib/web_system/web/panels/request_panel_test.rb
@@ -1,0 +1,14 @@
+class WebSystem
+  class RequestPanelTest < Liza::PanelTest
+
+    test :settings do
+      assert subject_class.log_level == :normal
+      assert subject_class.log_color == :blue
+    end
+
+    # test :call do
+    #   todo "write this"
+    # end
+
+  end
+end